rename : signup,signin CSS file rename AuthPage
